v2.2.0

+++ BACKEND/app/src/main/kotlin/_4Tipsy/TinyCloudAPI/routing/DownloadRouting.kt
@@ -4,11 +4,12 @@ package _4Tipsy.TinyCloudAPI.routing
 
 import io.ktor.server.routing.*
 import io.ktor.server.response.*
-import io.ktor.http.HttpStatusCode
-import io.ktor.http.HttpHeaders
 import io.ktor.server.application.call
 import io.ktor.server.application.install
 import io.ktor.server.plugins.partialcontent.PartialContent
+import io.ktor.http.HttpStatusCode
+import io.ktor.http.HttpHeaders
+import io.ktor.http.quote // 0_o
 
 import java.nio.file.Files
 
@@ -17,11 +18,6 @@ import _4Tipsy.TinyCloudAPI.services.DownloadService
 import _4Tipsy.TinyCloudAPI.guards.AuthGuard
 import _4Tipsy.TinyCloudAPI.guards.GetValidEid
 import _4Tipsy.TinyCloudAPI.exceptions.HttpException
-
-
-
-
-
 
 
 
@@ -55,17 +51,48 @@ fun Routing.downloadRouting() {
         )
 
         // if ok
-        call.response.headers.append(HttpHeaders.ContentDisposition, "attachment; filename=\"$fileName\"")
-        call.respondFile(file)
-
-        // cleanup (if it was retrieved dir)
-        if (file.name.endsWith(".tmp")) Files.delete(file.toPath()) // `file.name`, not `fileName`
+        try {
+          call.response.headers.append(HttpHeaders.ContentDisposition, "attachment; filename=${fileName.quote()}")
+          call.respondFile(file)
+        } finally {
+          // cleanup (if it was retrieved dir) // in `finally` cuz call might be interrupted
+          if (file.name.endsWith(".tmp")) Files.delete(file.toPath()) // `file.name`, not `fileName`
+        }
       }
     }
 
 
 
+  route("/download-by-eid") {
 
+    install(PartialContent) // set as streaming
+
+
+    // handler
+    get {
+      val targetEid = call.parameters.get("target")
+        ?: throw HttpException(HttpStatusCode.UnprocessableEntity, "Request validation error", "No 'target' param provided")
+
+      val uid = AuthGuard(call)
+
+
+      if (targetEid == "null") throw HttpException(HttpStatusCode(418, "I'm a teapot"), "Root has no EID, this route not supposed to download it", "Use regular '/download', silly ;3")
+      val (file, fileName) = DownloadService.downloadEntity(
+        target = "@eid="+targetEid,
+        _targetEid = targetEid,
+        uid = uid
+      )
+
+      // if ok
+      try {
+        call.response.headers.append(HttpHeaders.ContentDisposition, "attachment; filename=${fileName.quote()}")
+        call.respondFile(file)
+      } finally {
+        // cleanup (if it was retrieved dir) // in `finally` cuz call might be interrupted
+        if (file.name.endsWith(".tmp")) Files.delete(file.toPath()) // `file.name`, not `fileName`
+      }
+    }
+  }
